+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pantheon-systems/pcc-cli
 
+## 5.1.1-beta.1
+
+### Patch Changes
+
+- Updated dependencies [302023e]
+  - @pantheon-systems/pcc-sdk-core@5.1.1-beta.1
+
+## 5.1.1-beta.0
+
+### Patch Changes
+
+- 2bab8bf: Show --appRouter in possible flags for init command.
+  - @pantheon-systems/pcc-sdk-core@5.1.1-beta.0
+
 ## 5.1.0
 
 ### Minor Changes