+++ packages/react-sdk/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pantheon-systems/pcc-react-sdk
 
+## 5.1.1-beta.1
+
+### Patch Changes
+
+- 302023e: Fixed 404 redirect issue with document preview and publish
+- Updated dependencies [302023e]
+  - @pantheon-systems/pcc-sdk-core@5.1.1-beta.1
+
+## 5.1.1-beta.0
+
+### Patch Changes
+
+- 28c7e14: Remove version info from status endpoint.
+  - @pantheon-systems/pcc-sdk-core@5.1.1-beta.0
+
 ## 5.1.0
 
 ### Minor Changes